Moved Macros logic into MacroImplicits
